requirements.txt
openai>=1.0.0
openai-whisper>=20231117
requests>=2.31.0
pygame>=2.5.2
python-dotenv>=1.0.0
sounddevice
soundfile
PyYAML>=6.0
# PyTorch CUDA 11.8 stack (works with your RTX 3050)
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
pydub
transformers>=4.30.0
pyannote.audio==3.3.2
numpy==1.26.4
